move everything to different targets
