+++ Package.swift
@@ -34,7 +34,59 @@ let package = Package(
                 .product(name: "AsyncHTTPClient", package: "async-http-client"),
                 .product(name: "Atomics", package: "swift-atomics"),
                 .product(name: "MultipartKit", package: "multipart-kit"),
-                "WebSocketKitFork"
+                "DiscordAuth",
+                "DiscordClient",
+                "DiscordCore",
+                "DiscordGateway",
+                "DiscordLogger",
+                "DiscordModels",
+                "DiscordUtils",
+            ]
+        ),
+        .target(
+            name: "DiscordClient",
+            dependencies: [
+                .product(name: "AsyncHTTPClient", package: "async-http-client"),
+                "DiscordModels",
+            ]
+        ),
+        .target(
+            name: "DiscordCore",
+            dependencies: [
+                .product(name: "Logging", package: "swift-log"),
+                .product(name: "MultipartKit", package: "multipart-kit"),
+            ]
+        ),
+        .target(
+            name: "DiscordGateway",
+            dependencies: [
+                .product(name: "NIOCore", package: "swift-nio"),
+                .product(name: "AsyncHTTPClient", package: "async-http-client"),
+                "WebSocketKitFork",
+                "DiscordClient",
+            ]
+        ),
+        .target(
+            name: "DiscordLogger",
+            dependencies: [
+                .product(name: "NIOCore", package: "swift-nio"),
+                .product(name: "Logging", package: "swift-log"),
+                "DiscordClient",
+                "DiscordUtils",
+            ]
+        ),
+        .target(
+            name: "DiscordModels",
+            dependencies: [
+                .product(name: "MultipartKit", package: "multipart-kit"),
+                "DiscordCore"
+            ]
+        ),
+        .target(name: "DiscordUtils"),
+        .target(
+            name: "DiscordAuth",
+            dependencies: [
+                "DiscordModels"
             ]
         ),
         /// `WebSocketKit` will be replaced as soon as changes are final and merged in
