+++ Sources/DiscordAuth/BotAuthManager.swift
@@ -1,3 +1,4 @@
+import DiscordModels
 
 private let baseURLs = (
     authorization: "https://discord.com/api/oauth2/authorize",
