+++ Sources/DiscordBM/exports.swift
@@ -1,0 +1,7 @@
+@_exported import DiscordAuth
+@_exported import DiscordClient
+@_exported import DiscordCore
+@_exported import DiscordGateway
+@_exported import DiscordLogger
+@_exported import DiscordModels
+@_exported import DiscordUtils
