+++ Sources/DiscordCore/+Array.swift
@@ -1,6 +1,6 @@
 
 extension Array where Element == (String, String?) {
-    func makeForURLQuery() -> String {
+    public func makeForURLQuery() -> String {
         self.compactMap { key, value -> (String, String)? in
             guard let value = value else { return nil }
             return (key, value)
@@ -9,7 +9,7 @@ extension Array where Element == (String, String?) {
 }
 
 extension Array where Element == (String, String) {
-    func makeForURLQuery() -> String {
+    public func makeForURLQuery() -> String {
         if self.isEmpty {
             return ""
         } else {
