+++ Sources/DiscordCore/DiscordGlobalConfiguration.swift
@@ -29,19 +29,6 @@ public enum DiscordGlobalConfiguration {
     public static var performClientValidations = true
 }
 
-//MARK: - Internal DiscordGlobalConfiguration
-extension DiscordGlobalConfiguration {
-    static func makeDecodeLogger(_ label: String) -> Logger {
-        if enableLoggingDuringDecode {
-            var logger = DiscordGlobalConfiguration.makeLogger(label)
-            logger[metadataKey: "explanation"] = "If you're using one of the recent versions of DiscordBM, please report this on https://github.com/MahdiBM/DiscordBM/issues if there are no similar issues, so we can keep DiscordBM up to date for the community. You can disable these logs by using 'DiscordGlobalConfiguration.enableLoggingDuringDecode = false'"
-            return logger
-        } else {
-            return Logger(label: label, factory: SwiftLogNoOpLogHandler.init)
-        }
-    }
-}
-
 //MARK: - DiscordDecoder
 public protocol DiscordDecoder {
     func decode<D: Decodable>(_ type: D.Type, from: Data) throws -> D
