+++ Sources/DiscordGateway/GatewayManager.swift
