+++ Sources/DiscordGateway/SerialQueue.swift
