+++ Sources/DiscordGateway/exports.swift
@@ -1,0 +1,1 @@
+@_exported import DiscordClient
