+++ Sources/DiscordModels/Types/Auto Moderation.swift
