+++ Sources/DiscordModels/Types/Emoji.swift
