+++ Sources/DiscordModels/Types/Guild Scheduled Event.swift
