+++ Sources/DiscordModels/Types/Integration.swift
