+++ Sources/DiscordModels/Types/Interaction.swift
