+++ Sources/DiscordModels/Types/Permissions.swift
