+++ Sources/DiscordModels/Types/Sticker.swift
