+++ Sources/DiscordModels/Utils/Validatable.swift
