+++ Tests/DiscordBMTests/Models.swift
@@ -1,8 +1,8 @@
-@testable import DiscordBM
+@testable import DiscordModels
 import NIOCore
 import XCTest
 
-class GatewayModelsTests: XCTestCase {
+class ModelsTests: XCTestCase {
     
     func testEventDecode() throws {
         
@@ -244,6 +244,30 @@ class GatewayModelsTests: XCTestCase {
             XCTAssertEqual(file?.data, data)
             XCTAssertEqual(file?.extension, "png")
         }
+    }
+    
+    func testWebhookAddress() throws {
+        let webhookUrl = "https://discord.com/api/webhooks/1066287437724266536/dSmCyqTEGP1lBnpWJAVU-CgQy4s3GRXpzKIeHs0ApHm62FngQZPn7kgaOyaiZe6E5wl_"
+        let expectedId = "1066287437724266536"
+        let expectedToken = "dSmCyqTEGP1lBnpWJAVU-CgQy4s3GRXpzKIeHs0ApHm62FngQZPn7kgaOyaiZe6E5wl_"
+        
+        let (id1, token1) = try XCTUnwrap(
+            WebhookAddress.url(webhookUrl).deconstruct()
+        )
+        XCTAssertEqual(id1, expectedId)
+        XCTAssertEqual(token1, expectedToken)
+        
+        let (id2, token2) = try XCTUnwrap(
+            WebhookAddress.url(webhookUrl + "/").deconstruct()
+        )
+        XCTAssertEqual(id2, expectedId)
+        XCTAssertEqual(token2, expectedToken)
+        
+        let (id3, token3) = try XCTUnwrap(
+            WebhookAddress.url(webhookUrl).deconstruct()
+        )
+        XCTAssertEqual(id3, expectedId)
+        XCTAssertEqual(token3, expectedToken)
     }
 }
