+++ Tests/DiscordBMTests/SerialQueue.swift
@@ -1,4 +1,4 @@
-@testable import DiscordBM
+@testable import DiscordGateway
 import Atomics
 import XCTest
